PYTHON_chs_lib/spikeflag.py modified: spikeflag.py bug fix
git-svn-id: https://svn.ufz.de/svn/chs-svn/PYTHON_chs_lib@1722 20e3f476-f421-4892-ba25-73c5d2d2c40b

+++ spikeflag.py
@@ -80,7 +80,7 @@ def spikeflag(date, data, inflag, isday, window=13, iter=1,
     Written,  AP, Aug 2014
     '''       
     rows, cols = np.shape(data)
-    flag       = np.zeros_like(inflag)
+    flag       = np.zeros_like(inflag).astype(np.int)
     
     # mad window length and flag window length 
     period   = np.int(window*t_int)/2
